node:http: move the subprocess exit test next to the other spawn tests; bind the proxy test to 127.0.0.1
The abort-events file's scenarios are verified against Node.js, so keep it free
of Bun-only helpers. The proxy test's listen("localhost") can bind ::1 while
the client connects to 127.0.0.1; bind the IPv4 loopback explicitly.
